requirements.txt
alembic==1.0.5
aniso8601==4.0.1
bcrypt==3.1.5
blinker==1.4
cffi==1.11.5
Click==7.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.3.1
flask-restplus==0.12.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
Flask-WhooshAlchemy==0.56
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==2.6.0
Mako==1.0.7
MarkupSafe==1.1.0
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.7.5
python-editor==1.0.3
pytz==2018.7
six==1.12.0
SQLAlchemy==1.2.15
Werkzeug==0.14.1
Whoosh==2.7.4
WTForms==2.2.1
flask-httpauth-3.2.4
